move common from app to src
https://coveord.atlassian.net/browse/KIT-2708
