+++ packages/samples/headless-ssr/src/app/generic/common/engine.ts
@@ -3,7 +3,7 @@ import {
   InferStaticState,
   InferHydratedState,
 } from '@coveo/headless/ssr';
-import {config} from '../../common/search-engine-config';
+import {config} from '../../../common/search-engine-config';
 
 const engineDefinition = defineSearchEngine(config);
