+++ packages/samples/headless-ssr/src/app/react/common/engine.ts
@@ -1,6 +1,6 @@
 import {defineSearchEngine} from '@coveo/headless-react/ssr';
 import {InferStaticState, InferHydratedState} from '@coveo/headless/ssr';
-import {config} from '../../common/search-engine-config';
+import {config} from '../../../common/search-engine-config';
 
 const engineDefinition = defineSearchEngine(config);
