+++ packages/samples/headless-ssr/src/app/react/components/facets.tsx
@@ -1,6 +1,6 @@
 import {FacetState, Facet as FacetController} from '@coveo/headless/ssr';
 import {useEffect, useState, FunctionComponent} from 'react';
-import FacetCommon from '../../common/facet';
+import FacetCommon from '../../../common/facet';
 import {useAuthorFacet} from '../common/engine';
 
 export const AuthorFacet = () => {
