+++ packages/samples/headless-ssr/src/app/react/components/result-list.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import ResultListCommon from '../../common/result-list';
+import ResultListCommon from '../../../common/result-list';
 import {useResultList} from '../common/engine';
 
 export default function ResultList() {
