+++ packages/samples/headless-ssr/src/app/react/components/search-box.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import SearchBoxCommon from '../../common/search-box';
+import SearchBoxCommon from '../../../common/search-box';
 import {useSearchBox} from '../common/engine';
 
 export default function SearchBox() {
