+++ packages/samples/headless-ssr/src/common/facet-search.tsx
