+++ packages/samples/headless-ssr/src/common/facet.tsx
