+++ packages/samples/headless-ssr/src/common/hydration-metadata.tsx
@@ -1,5 +1,8 @@
 import {FunctionComponent} from 'react';
-import {SearchHydratedState, SearchStaticState} from '../generic/common/engine';
+import {
+  SearchHydratedState,
+  SearchStaticState,
+} from '../app/generic/common/engine';
 
 export interface HydrationMetadataProps {
   staticState: SearchStaticState;
