+++ packages/samples/headless-ssr/src/common/result-list.tsx
