+++ packages/samples/headless-ssr/src/common/search-box.tsx
