+++ packages/samples/headless-ssr/src/common/search-engine-config.ts
