+++ packages/samples/headless-ssr/src/common/search-parameters-serializer.ts
